Make ipListIntervalTree private to RDAPControllerUtil
